+++ substratevm/CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog summarizes major changes to GraalVM Native Image.
 * (GR-70136) Add a new tool `--tool:llvm` for the LLVM backend of Native Image.
 * (GR-68984) Ship the `reachability-metadata-schema.json` together with GraalVM at `<graalvm-home>/lib/svm/schemas/reachability-metadata-schema.json`.
 * (GR-68984) Improve the schema to capture detailed constraints about each element in the `reachability-metadata-schema.json`.
+* (GR-70046) Remove all support for running image builder on classpath.
 
 ## GraalVM 25
 * (GR-52276) (GR-61959) Add support for Arena.ofShared().